requirements.txt
matplotlib>=3.3.3
numpy>=1.18.5
pandas>=0.25.3
Pillow>=8.4.0
scikit_learn>=1.0.1
torch>=1.1.0
torchvision>=0.3.0
tqdm>=4.54.0
